remove unused code

+++ bestla/bestla/sycl/sycl_prologue_b.h
@@ -318,7 +318,7 @@ class WeightS4Trans {
     int nsg_k = k / GroupK;
     int nsg_n = n / GroupN;
     sycl::range<1> group{SgSize};
-    sycl::range<1> problem{nsg_n * nsg_k * SgSize};
+    sycl::range<1> problem{static_cast<size_t>(nsg_n) * nsg_k * SgSize};
     auto B_d = in.B;
     auto S_d = in.scale;
     int ldb = in.ldb;
@@ -371,7 +371,7 @@ class WeightS4Trans {
     int constexpr Unroll = 2;
     int constexpr SgSize = 16;
     sycl::range<1> group{SgSize};
-    sycl::range<1> problem{n * SgSize};
+    sycl::range<1> problem{static_cast<size_t>(n) * SgSize};
     if (k % (SgSize * 32 * Unroll) == 0) {
       int constexpr TileK = 32;
       int constexpr GroupK = SgSize * TileK;
